Refactor: Initial project restructuring and modularization
This commit marks the completion of the initial refactoring phase for the Scholarino bot.

Key changes include:
- **Modular Architecture:** The monolithic `main.py` has been broken down into a structured and maintainable architecture under the `src/` directory. Code is now organized into handlers, services, data, and utils, improving separation of concerns.
- **Dependency Management:** A comprehensive `requirements.txt` file has been created to ensure a reproducible development environment.
- **Improved Documentation:** The `README.md` has been completely rewritten to provide a clear overview of the project, its architecture, setup instructions, and future roadmap.
- **Configuration Management:** A dedicated `src/config.py` module now handles environment variables and centralizes configuration.
- **Code Refinements:**
  - Google Sheets service now loads credentials from environment variables directly, which is better for cloud deployments.
  - The main `bot.py` script is now cleaner and acts as an orchestrator, registering handlers from their respective modules.
  - Added support for local development via polling.
