+++ README.md
@@ -1,2 +1,112 @@
-# smartbot
-robot smart for scholarships in perugia 
+# Scholarino Bot 🇮🇹🤖
+
+**ربات هوشمند راهنمای دانشجویان ایرانی و بین‌المللی در پروجا، ایتالیا**
+
+[![زبان](https://img.shields.io/badge/language-Python-blue.svg)](https://www.python.org/)
+[![کتابخانه](https://img.shields.io/badge/library-python--telegram--bot-green.svg)](https://python-telegram-bot.org/)
+[![پلتفرم](https://img.shields.io/badge/platform-Telegram-blue.svg)](https://telegram.org/)
+
+---
+
+## 🎯 درباره پروژه
+
+**Scholarino** یک ربات تلگرامی است که با هدف ارائه راهنمایی‌های دقیق، شخصی‌سازی‌شده و هوشمند به دانشجویان ایرانی و بین‌المللی در شهر پروجا طراحی شده است. این ربات به سوالات متداول در مورد بورسیه استانی (ADiSU)، مراحل اداری پس از ورود، زندگی روزمره، و فرصت‌های دانشجویی پاسخ می‌دهد.
+
+### ✨ ویژگی‌های کلیدی
+
+- **پاسخگویی هوشمند:** با استفاده از **OpenAI (GPT-4o)**، ربات به سوالات متنی کاربران پاسخ‌های دقیق و مرتبط می‌دهد.
+- **پایگاه دانش جامع:** اطلاعات ساختاریافته در مورد موضوعات مختلف از طریق یک فایل `knowledge.json` مدیریت می‌شود که به راحتی قابل به‌روزرسانی است.
+- **چندزبانه (در حال توسعه):** معماری پروژه برای پشتیبانی از زبان‌های فارسی، انگلیسی و ایتالیایی طراحی شده است.
+- **شخصی‌سازی:** ربات اطلاعات اولیه کاربر (نام، سن) را برای ارائه پاسخ‌های دوستانه‌تر ذخیره می‌کند.
+- **یکپارچه‌سازی با Google Sheets:** برای ذخیره‌سازی پروفایل کاربران و تاریخچه پرسش و پاسخ‌ها (در نسخه اولیه).
+- **معماری ماژولار و مقیاس‌پذیر:** کد به بخش‌های منطقی تقسیم شده تا توسعه و نگهداری آن آسان باشد.
+
+---
+
+## 🏛️ معماری پروژه
+
+پروژه با پیروی از اصول کد تمیز و ماژولار بازسازی شده است:
+
+- **`src/`**: قلب برنامه که شامل تمام منطق ربات است.
+  - **`handlers/`**: ماژول‌های مدیریت‌کننده انواع تعاملات کاربر (شروع، منو، پیام متنی).
+  - **`services/`**: ماژول‌های ارتباط با سرویس‌های خارجی مانند OpenAI و Google Sheets.
+  - **`data/`**: محل نگهداری پایگاه دانش (`knowledge.json`).
+  - **`utils/`**: ابزارهای کمکی مانند سازنده کیبورد.
+  - **`config.py`**: مدیریت تنظیمات و متغیرهای محیطی.
+  - **`bot.py`**: نقطه ورود اصلی برنامه که ربات را راه‌اندازی و گرداننده‌ها را ثبت می‌کند.
+- **`requirements.txt`**: لیست تمام وابستگی‌های پروژه.
+- **`Dockerfile` (در آینده):** برای ساخت ایمیج داکر و استقرار آسان.
+- **`docs/`**: مستندات تکمیلی پروژه.
+
+---
+
+## 🚀 راه‌اندازی و نصب
+
+برای اجرای پروژه به صورت محلی، مراحل زیر را دنبال کنید:
+
+### ۱. پیش‌نیازها
+
+- [پایتون](https://www.python.org/downloads/) نسخه 3.9 یا بالاتر
+- حساب کاربری تلگرام و [توکن ربات](https://core.telegram.org/bots#6-botfather)
+- [کلید API از OpenAI](https://platform.openai.com/api-keys)
+- دسترسی به [Google Sheets API](https://docs.gspread.org/en/latest/oauth2.html) و شناسه شیت (`SHEET_ID`).
+
+### ۲. نصب وابستگی‌ها
+
+ابتدا پروژه را کلون کنید:
+```bash
+git clone https://github.com/your-username/scholarino-bot.git
+cd scholarino-bot
+```
+
+سپس یک محیط مجازی ایجاد کرده و وابستگی‌ها را نصب کنید:
+```bash
+python -m venv venv
+source venv/bin/activate  # On Windows use `venv\Scripts\activate`
+pip install -r requirements.txt
+```
+
+### ۳. تنظیم متغیرهای محیطی
+
+یک فایل `.env` در ریشه پروژه ایجاد کرده و متغیرهای زیر را در آن قرار دهید:
+
+```env
+TELEGRAM_TOKEN="YOUR_TELEGRAM_BOT_TOKEN"
+OPENAI_API_KEY="YOUR_OPENAI_API_KEY"
+SHEET_ID="YOUR_GOOGLE_SHEET_ID"
+ADMIN_CHAT_ID="YOUR_TELEGRAM_CHAT_ID" # برای دریافت نوتیفیکیشن‌ها
+
+# محتوای فایل JSON اعتبارنامه Google را به صورت یک خطی اینجا کپی کنید
+GOOGLE_CREDS_JSON='{"type": "service_account", "project_id": "...", ...}'
+
+# برای استقرار با وب‌هوک
+# BASE_URL="https://your-app-name.onrender.com"
+# WEBHOOK_SECRET="A_STRONG_RANDOM_SECRET_STRING"
+```
+
+### ۴. اجرای ربات
+
+برای اجرای ربات در حالت توسعه (polling)، دستور زیر را اجرا کنید:
+```bash
+python src/bot.py
+```
+
+ربات شما اکنون فعال است و به پیام‌ها در تلگرام پاسخ می‌دهد.
+
+---
+
+## 🤝 مشارکت در پروژه
+
+مشارکت شما در توسعه Scholarino بسیار ارزشمند است! اگر ایده‌ای برای بهبود دارید یا با خطایی مواجه شدید، لطفاً یک **Issue** جدید ثبت کنید یا یک **Pull Request** ارسال نمایید.
+
+### نقشه راه آینده
+
+- [ ] انتقال کامل به پایگاه داده **PostgreSQL**.
+- [ ] پیاده‌سازی کامل سیستم **چندزبانگی (i18n)**.
+- [ ] افزودن **تست‌های خودکار** (Unit & Integration).
+- [ ] ایجاد **داشبورد ادمین** برای مدیریت کاربران و محتوا.
+- [ ] توسعه سیستم **نوتیفیکیشن هوشمند** بر اساس تقویم.
+
+---
+
+با تشکر از توجه شما! امیدواریم این ربات بتواند به دانشجویان عزیز در پروجا کمک کند.
