+++ requirements.txt
@@ -1,7 +1,7 @@
-python-telegram-bot[webhooks]==20.3
-gspread==5.7.0
-oauth2client==4.1.3
-python-dotenv==0.21.0
-openai==1.11.1
-tenacity==8.2.3
-functions-framework==3.4.0
+python-telegram-bot[job-queue, webhooks]>=20.8
+openai>=1.3
+gspread>=5.12
+oauth2client>=4.1
+Flask>=3.0
+functions-framework>=3.4
+requests>=2.31
